refactor #1 modularized command type def
