+++ src/pages/core/command/index.ts
@@ -1,0 +1,1 @@
+export type { Command } from "./types";
