+++ src/pages/core/command/types.ts
@@ -1,0 +1,10 @@
+export type Command = {
+  title: string;
+  subtitle?: string;
+  shortcut?: string;
+  lastVisitTime?: number;
+  keyword?: string;
+  icon?: string;
+  command?: () => unknown;
+  url?: string;
+};
