+++ src/pages/popup/commands/audio.ts
@@ -1,7 +1,8 @@
+import type { Command } from "@pages/core/command";
+
 import { createLazyResource, parsedInput } from "~/util/signals";
 
 import { faviconURL } from "../Entry";
-import { Command } from "./general";
 
 const baseCommands = [
   {
