+++ src/pages/popup/commands/bookmark-this.ts
@@ -1,7 +1,8 @@
+import type { Command } from "@pages/core/command";
+
 import { createLazyResource, matchCommand, setInput } from "~/util/signals";
 
 import { faviconURL } from "../Entry";
-import { Command } from "./general";
 
 const KEYWORD = "bt";
