+++ src/pages/popup/commands/themes.ts
@@ -1,8 +1,7 @@
+import type { Command } from "@pages/core/command";
 import { createEffect, createMemo, createSignal } from "solid-js";
 
 import { createStoredSignal, matchCommand, setInput } from "~/util/signals";
-
-import { Command } from "./general";
 
 const KEYWORD = "theme";
