+++ src/pages/popup/commands/website-search.ts
@@ -1,7 +1,8 @@
+import type { Command } from "@pages/core/command";
+
 import { matchCommand, parsedInput, setInput } from "~/util/signals";
 
 import { faviconURL } from "../Entry";
-import { Command } from "./general";
 
 type Template = {
   title: string;
